refactor(editor): extract focus mode button + header-footer clicks to own modules
Keeps editor.ts under the 200 effective-line ESLint limit.
